moved upsert_validator txs to transactions, updated imports
