+++ planetmint/__init__.py
@@ -4,7 +4,7 @@
 # Code is Apache-2.0 and docs are CC-BY-4.0
 
 from planetmint.transactions.common.transaction import Transaction  # noqa
-from planetmint.upsert_validator import ValidatorElection  # noqa
+from planetmint.transactions.types.elections.validator_election import ValidatorElection  # noqa
 from planetmint.transactions.types.elections.vote import Vote  # noqa
 from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
 from planetmint.lib import Planetmint
