--- planetmint/upsert_validator/__init__.py
@@ -1,7 +1,0 @@
-# Copyright © 2020 Interplanetary Database Association e.V.,
-# Planetmint and IPDB software contributors.
-# SPDX-License-Identifier: (Apache-2.0 AND CC-BY-4.0)
-# Code is Apache-2.0 and docs are CC-BY-4.0
-
-
-from planetmint.upsert_validator.validator_election import ValidatorElection  # noqa
