+++ tests/elections/test_election.py
@@ -5,7 +5,7 @@ from tests.utils import generate_election, generate_validators
 from planetmint.lib import Block
 from planetmint.transactions.types.elections.election import Election
 from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
-from planetmint.upsert_validator.validator_election import ValidatorElection
+from planetmint.transactions.types.elections.validator_election import ValidatorElection
 
 
 @pytest.mark.bdb
