+++ tests/tendermint/test_core.py
@@ -19,8 +19,8 @@ from planetmint.core import OkCode, CodeTypeError, rollback
 from planetmint.transactions.types.elections.election import Election
 from planetmint.lib import Block
 from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
-from planetmint.upsert_validator.validator_election import ValidatorElection
-from planetmint.upsert_validator.validator_utils import new_validator_set
+from planetmint.transactions.types.elections.validator_election import ValidatorElection
+from planetmint.transactions.types.elections.validator_utils import new_validator_set
 from planetmint.tendermint_utils import public_key_to_base64
 from planetmint.version import __tm_supported_versions__
