+++ tests/upsert_validator/test_validator_election.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 import pytest
 
 from planetmint.tendermint_utils import public_key_to_base64
-from planetmint.upsert_validator import ValidatorElection
+from planetmint.transactions.types.elections.validator_election import ValidatorElection
 from planetmint.transactions.common.exceptions import (
     DuplicateTransaction,
     UnequalValidatorSet,
